pdf_summary/pyproject.toml
[project]
name = "pdf-summary"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.2",
    "google-genai>=1.50.1",
    "httpx>=0.28.1",
    "mcp[cli]>=1.21.1",
    "uvicorn[standard]>=0.38.0",
]
